requirements.txt
asgiref==3.2.5
Django==3.0.4
django-redis==4.11.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
pkg-resources==0.0.0
PyJWT==1.7.1
pytz==2019.3
redis==3.4.1
sqlparse==0.3.1
